+++ app/Http/Controllers/API/V1/APIRemoveProviderLogoController.php
@@ -9,6 +9,7 @@ use App\Services\LogoService;
 use App\Models\Tenants\Provider;
 use Illuminate\Support\Facades\DB;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
 use Barryvdh\Debugbar\Facades\Debugbar;
 use Illuminate\Support\Facades\Storage;
 use App\Http\Requests\Tenant\ProviderRequest;
@@ -21,6 +22,9 @@ class APIRemoveProviderLogoController extends Controller
 
     public function destroy(Provider $provider)
     {
+        if (Auth::user()->cannot('update', $provider))
+            return ApiResponse::notAuthorized();
+        
         $this->logoService->deleteExistingFiles($provider);
         return ApiResponse::success('', 'Logo deleted');
     }
